correções

+++ src/main/java/gui/vieira/mecurso/repository/CursoRepository.java
@@ -5,6 +5,6 @@ import gui.vieira.mecurso.entity.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface CursoRepository extends JpaRepository<Curso, Float> {
+public interface CursoRepository extends JpaRepository<Curso, Long> {
     
 }
